DayTwo
Write a program that displays the following diamond shape (User input),
Write a program that calculates world population growth each year for
the next 75 years (Sample output shown below), using the assumption that
the current growth rate will stay constant, Modify existing class
definition provided as below:

1. Add two more fields to Hotel class to store the year it was built and
the star rating.

2. Modify constructor to initialize year and star rating.

3. Create set and get methods for year and star rating

4. Write a console C# application to test Hotel class. Inside the main
program validate user input for Rooms and Cost (rooms :10 to 100 and
cost: 50 to 500) before creating an object of hotel class, Consider
that, when creating a program for a real estate company that sells
houses, you want the program to ask a customer the type of house that he
or she wants to purchase and/or the type of garage that the desired
house should have. Here is an example of running the program:
Enter the type of house you want to purchase
1 - Single Family
2 - Townhouse
3 – Lifestyle block
Your Choice: 1

Enter the type of garage you want
0 - single
1 - Double
2 – 3 Car +
3 – Carport
4 – Not essential
Your Choice: 4